refactor(replay): Move HTTP capture impl under sentry_flutter
Network capture for replay is a Flutter-only concern; sentry (dart)
now only exposes an internal NetworkDetailsCapture interface plus a
nullable hook on SentryOptions, mirroring the Transport seam.
